requirements.txt
streamlit>=1.32
langgraph>=0.2.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
langsmith>=0.1.0
chromadb>=0.5.0
google-generativeai>=0.7.0
pydantic>=2.6
python-dotenv>=1.0
feedparser>=6.0
httpx>=0.27
beautifulsoup4>=4.12
fastapi>=0.110
uvicorn[standard]>=0.29
